A mapreduce v2 pipes driver.
First attempt to write a java side pipes driver based on mapreduce
v2. The main motivation is to be able to use 'modern' InputFormat(s),
e.g., Parquet. The latter are based on org.apache.hadoop.mapreduce and
not org.apache.hadoop.mapred,  and are thus refused with disdain by the
current implementation.

The code is basically a coarse adaptation of org.apache.mapred.pipes.