Fixed a timing bug in the network simulator.
From time to time, the network simulator was hanging depending on what,
between the simulator and the client program, was the first to start
waiting for data on the socket. Now it is guaranteed that the simulator
cmd_flux will start writing before reading, thus client has reading
precedence. This appears to have resolved the problem.