Added support for NullWritable keys and values.
The convention is that NullWritable values (usually directly null) will
be mapped to a zero length string.  It is up to the python client to
decide what to do with them.

There is no explicit unit test for this. Examples/parquet/py contains an
example application that explicitly uses this feature.