A composition recipe for a full apache 2.6.0 hadoop system.
The recipe uses shared volume for service logging and to share sw
deployment in /usr/local between the client and nodemanager(s) nodes.
Simply invoking docker-compose is not sufficient to get the system up,
since it requires the nodes to known each other ip address and
docker-compose does not provide a dns service. To compensate for that, I
am adding a script, share_etc_hosts.py, that updates the /etc/hosts
files of all nodes in the cluster.  The script 'start_all.sh' is a first
attempt at simplifying cluster startup.