Handle compiled scripts; use NamedTemporaryFile to handle and delete zip
This commit resolves an issue with python script that arose when calling
byte-compiled python modules.  The code now does not hard-code the .py
extension and instead uses the same extension that's on the module name.

The commit also changes the code to use NamedTemporaryFile to create and
delete the temporary zip archive with the code to be executed.  This
change results in pydoop creating the archive in a proper temp
directory, and NamedTemporaryFile should be more robust in cleaning up
the file once execution terminates (especially in case of error).