updated setup.py, and related sundries, to be python 3 friendly
Changed setup.py so that it can run both under 2.7 and >=3.5. Now it can attempt
build and miserably fail on C/C++ compilations.
Introduced a new dependency, on 'future', for python2.7.
We now state that we support both 2.7 and 3.5.

Few compatibility changes, all localized on python files that are loaded by
setup.py.