Patched a exception inconsistency between py2.7 and py3.5.
For some mysterious reason, local_file.write will raise a ValueError in py2.7
and a IOError in py3.5 when asked to write in a open for reading file. It is
probably some side effect of future. For the time being, I just capture the
ValueError and convert in a IOError.